requirements.txt
# Core dependencies
# No google-genai or openai needed - uses scaile-services OpenRouter gateway
pydantic>=2.0.0
langdetect>=1.0.9  # For language validation
aiohttp>=3.9.0
python-dotenv>=1.0.0
httpx>=0.25.0  # For HTTP requests to scaile-services
defusedxml>=0.7.1  # For secure XML parsing (XXE protection)
tenacity>=8.2.3  # For retry logic
fastapi>=0.104.0  # For REST API
uvicorn>=0.24.0  # For ASGI server

# Google APIs (for Drive upload in image generation - NOT for AI)
google-api-python-client>=2.100.0
google-auth>=2.23.0

# Optional dependencies
supabase>=2.0.0  # For Supabase storage
replicate>=0.25.0  # For image generation (legacy)
pillow>=10.0.0  # For image processing in Google Imagen

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.5.0

